This report, by its very length, defends itself against the risk of being read

+++ 0x05-pointers_arrays_strings/2-strlen.c
@@ -7,7 +7,7 @@
 
 void _strlen(char *s)
 {
-  size_t str_length;
+  int str_length;
 
   for(str_lenth = 0; s[str_length] != '\0'; ++str_length);
